added random respawn
It only works well if the chunk is pre-generated.
If its not then I have added small bad hotfix but it will invalidate the block check if the chunk was not pre-generated.

Known Issues :
Random respawn on always join is triggered auto even when config file is set to false.